Move from lib/ to src/
