+++ index.js
@@ -1,1 +1,1 @@
-export * from './lib'
+export * from './src/index'
