+++ src/helpers/babel_es5_adapter.js
